fix: Apply MantineProvider in main.tsx
- Use MantineProvider and theme object in main.tsx instead of App.tsx
