+++ .gitignore
@@ -25,4 +25,3 @@ dist-ssr
 
 # Evironments
 .env
-my-documentation.md
